refactor: validation in store
