+++ src/js/Stores/global/global.js
@@ -1,9 +1,13 @@
 import { ADD_CAR_NAMES, ITERATION_COUNT, RESET } from './actions.js';
-import { GlobalStore } from '../../Models/GlobalStore.js';
+import { GlobalStore } from './models/GlobalStore.js';
 
 let state = new GlobalStore();
 
-function setState(newState) {
+function setState(newState, action) {
+  if (!GlobalStore.validateProps(newState)) {
+    throw new Error(`Error on action of ${action} with new State : ${newState}`);
+  }
+
   state = new GlobalStore(newState);
 }
 
@@ -13,14 +17,14 @@ export function dispatch(action, payload) {
       setState({
         ...state,
         carNames: payload,
-      });
+      }, action);
       break;
     }
     case(ITERATION_COUNT): {
       setState({
         ...state,
         iterationCount: payload,
-      });
+      }, action);
       break;
     }
     case(RESET):
@@ -28,7 +32,7 @@ export function dispatch(action, payload) {
       setState({
         carNames: [],
         iterationCount: 0,
-      });
+      }, action);
     }
   }
 }
